pyproject.toml
[tool.poetry]
name = "url_shortner"
version = "0.1.0"
description = ""
authors = ["Mohfarrej Kin Tuda"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.70.0"
uvicorn = {extras = ["standard"], version = "^0.15.0"}
SQLAlchemy = "^1.4.27"
pydantic = {extras = ["dotenv"], version = "^1.8.2"}
alembic = "^1.7.5"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
psycopg2 = "^2.9.2"
FastAPI-SQLAlchemy = "^0.2.1"
python-dotenv = "^0.19.2"
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
